move loadandbuild to a separate file fixes #12
